Add timeSeriesWindowComponent to share logic between spatial and non spatial timeseriesdisplay's
