+++ src/components/timeseries/TimeSeriesWindowComponent.vue
@@ -1,0 +1,202 @@
+<template>
+  <WindowComponent :hideToolbar="settings.general.hideToolbar">
+    <template v-slot:toolbar>
+      <v-toolbar-items
+        class="flex-0-0"
+        v-model="displayType"
+        mandatory
+        size="small"
+      >
+        <v-btn
+          v-for="item in displayTypeItems"
+          :key="item.value"
+          :value="item.value"
+          :aria-label="item.label"
+          :text="item.label"
+          :active="displayType === item.value"
+          variant="text"
+          width="20px"
+          class="pa-0 text-capitalize"
+          @click="displayType = item.value"
+          :disabled="item.disabled"
+        >
+          <v-icon :style="item.iconStyle">{{ item.icon }}</v-icon>
+        </v-btn>
+      </v-toolbar-items>
+
+      <slot name="toolbar-title">
+        <span class="ml-5">{{ displayConfig?.title }}</span>
+      </slot>
+      <v-spacer />
+    </template>
+    <template v-slot:toolbar-append>
+      <v-btn size="small" icon v-if="displayActionItems.length">
+        <v-icon>mdi-dots-horizontal</v-icon>
+        <v-menu activator="parent" density="compact">
+          <v-list>
+            <v-list-item
+              v-for="item in displayActionItems"
+              :key="item.label"
+              @click="item.action"
+              :prepend-icon="item.icon"
+            >
+              {{ item.label }}
+            </v-list-item>
+          </v-list>
+        </v-menu>
+      </v-btn>
+      <slot name="toolbar-append" />
+    </template>
+    <TimeSeriesComponent
+      :config="displayConfig ?? undefined"
+      :elevation-chart-config="elevationChartDisplayconfig ?? undefined"
+      :current-time="currentTime"
+      :displayType="displayType"
+      :information-content="informationContent"
+      :settings="settings"
+    >
+    </TimeSeriesComponent>
+    <TimeSeriesFileDownloadComponent
+      :config="displayConfig"
+      :options="options"
+      :filter="filter"
+    >
+    </TimeSeriesFileDownloadComponent>
+  </WindowComponent>
+</template>
+
+<script setup lang="ts">
+import WindowComponent from '@/components/general/WindowComponent.vue'
+import TimeSeriesComponent from '@/components/timeseries/TimeSeriesComponent.vue'
+import TimeSeriesFileDownloadComponent from '@/components/download/TimeSeriesFileDownloadComponent.vue'
+import { DisplayConfig, DisplayType } from '@/lib/display/DisplayConfig'
+import { type ChartSettings } from '@/lib/topology/componentSettings'
+import { computed, ref, StyleValue, watch } from 'vue'
+import { useDownloadDialogStore } from '@/stores/downloadDialog'
+import { UseDisplayConfigOptions } from '@/services/useDisplayConfig'
+import { useUserSettingsStore } from '@/stores/userSettings'
+import type {
+  filterActionsFilter,
+  timeSeriesGridActionsFilter,
+} from '@deltares/fews-pi-requests'
+
+interface Props {
+  displayConfig?: DisplayConfig | null
+  elevationChartDisplayconfig?: DisplayConfig | null
+  currentTime?: Date
+  informationContent?: string | null
+  filter?: filterActionsFilter | timeSeriesGridActionsFilter
+  settings: ChartSettings
+}
+
+const props = defineProps<Props>()
+
+const downloadDialogStore = useDownloadDialogStore()
+const userSettings = useUserSettingsStore()
+
+const options = computed<UseDisplayConfigOptions>(() => {
+  return {
+    useDisplayUnits: userSettings.useDisplayUnits,
+    convertDatum: userSettings.convertDatum,
+  }
+})
+
+interface DisplayTypeItem {
+  icon: string
+  label: string
+  value: DisplayType
+  iconStyle?: StyleValue
+  disabled?: boolean
+}
+
+const displayActionItems = computed(() => {
+  const dataDownloadEnabled = props.settings.actions.downloadData
+  return [
+    {
+      icon: 'mdi-download',
+      label: 'Download time series ...',
+      action: () => {
+        downloadDialogStore.showDialog = true
+      },
+      disabled: downloadDialogStore.disabled,
+      hidden: !dataDownloadEnabled,
+    },
+  ].filter((item) => !item.hidden)
+})
+
+function toDisplayType(
+  value: typeof props.settings.general.startPanel,
+): DisplayType {
+  switch (value) {
+    case 'metaDataPanel':
+      return DisplayType.Information
+    case 'timeseriesChart':
+      return DisplayType.TimeSeriesChart
+    case 'timeseriesTable':
+      return DisplayType.TimeSeriesTable
+    case 'verticalProfileChart':
+      return DisplayType.ElevationChart
+    case 'verticalProfileTable':
+      throw new Error('Vertical profile table is not supported')
+  }
+}
+
+const displayType = ref<DisplayType>(
+  toDisplayType(props.settings.general.startPanel),
+)
+const displayTypeItems = computed<DisplayTypeItem[]>(() => {
+  const noElevationCharts = !(
+    (props.elevationChartDisplayconfig?.subplots?.length ?? 0) > 0
+  )
+  const elevationChartsDefined = props.elevationChartDisplayconfig !== undefined
+
+  const noTooltip = props.informationContent === null
+  const tooltipDefined = props.informationContent !== undefined
+
+  const chartEnabled = props.settings.timeseriesChart.show
+  const elevationChartEnabled =
+    props.settings.verticalProfileChart.show && elevationChartsDefined
+  const tableEnabled = props.settings.timeseriesTable.show
+  const metaDataEnabled = props.settings.metaDataPanel.show && tooltipDefined
+  return [
+    {
+      icon: 'mdi-chart-line',
+      label: 'Chart',
+      value: DisplayType.TimeSeriesChart,
+      hidden: !chartEnabled,
+    },
+    {
+      icon: 'mdi-elevation-rise',
+      label: 'Vertical profile',
+      value: DisplayType.ElevationChart,
+      iconStyle: 'transform: rotate(-90deg);',
+      disabled: noElevationCharts,
+      hidden: !elevationChartEnabled,
+    },
+    {
+      icon: 'mdi-table',
+      label: 'Table',
+      value: DisplayType.TimeSeriesTable,
+      hidden: !tableEnabled,
+    },
+    {
+      icon: 'mdi-information-outline',
+      label: 'Information',
+      value: DisplayType.Information,
+      disabled: noTooltip,
+      hidden: !metaDataEnabled,
+    },
+    // Mdi icon for metatdata at the current location
+    //    {
+    //     icon: 'mdi-information',
+    //    label: 'Metadata',
+  ].filter((item) => !item.hidden)
+})
+
+watch(displayTypeItems, () => {
+  const activeItems = displayTypeItems.value.map((dt) => dt.value)
+  if (!activeItems.includes(displayType.value)) {
+    displayType.value = activeItems[0]
+  }
+})
+</script>
